+++ __tests__/sendBox.js
@@ -26,7 +26,7 @@ test('should focus send box when message is being sent', async () => {
 
   await driver.wait(uiConnected(), timeouts.directLine);
   await pageObjects.sendMessageViaSendBox('Hello, World!', { waitForSend: true });
-  await driver.wait(minNumActivitiesShown(1), timeouts.directLine);
+  await driver.wait(minNumActivitiesShown(2), timeouts.directLine);
 
   const base64PNG = await driver.takeScreenshot();